Move to the SeleniumScreenshotTaker so that we get the borderstyle.
